Fix audit event author id member type

+++ src/GitLabApiClient/Models/AuditEvents/Responses/AuditEvent.cs
@@ -6,10 +6,10 @@ namespace GitLabApiClient.Models.AuditEvents.Responses;
 public class AuditEvent
 {
     [JsonProperty("id")]
-    public long Id { get; set; }
+    public int Id { get; set; }
 
     [JsonProperty("author_id")]
-    public long AuthorId { get; set; }
+    public int AuthorId { get; set; }
 
     [JsonProperty("entity_type")]
     public string EntityType { get; set; }
